Update README.md
It's an AI-based virtual mouse project that uses Python and OpenCV for real-time hand gesture tracking. It detects and tracks hand landmarks using Mediapipe and maps gestures to mouse actions. Autopy is utilized to control mouse movement and clicks. This project offers an alternative way of controlling the mouse without a physical device, enhancing user interaction with computers. It showcases the power of computer vision and AI in creating intuitive and hands-free interfaces.